fully functional, no functions

+++ userdata.c
@@ -6,7 +6,8 @@
 #include <stdio.h>
 
 void printFormHeader(char*);
-int paymentSum(int foodPrice, int drinkPrice, int drinkCode, int noDrink);
+
+double paymentSum(double foodPrice, double drinkPrice, int drinkCode, int noDrink);
 
 void signIn(char *username, char *password) {
     printf("Welcome to Food Thingies!\n");
@@ -19,13 +20,13 @@ void signIn(char *username, char *password) {
 
 void printForm(char* username, char* food, double foodPrice, char* drink, double drinkPrice, int drinkCode, int noDrink, int cutleryChoice, char *additionalInfo) {
     printFormHeader(username);
-    printf("--%s: %d\n", food, foodPrice);
+    printf("--%s: %f\n", food, foodPrice);
     if (drinkCode != noDrink) {
-        printf("--%s: %d\n", drink, drinkPrice);
+        printf("--%s: %f\n", drink, drinkPrice);
     }
     printf("Cutlery: %s\n", (cutleryChoice == 1? "no":"yes"));
     printf("Additional information: %s\n", additionalInfo);
-    printf("Payment amount: %d\n", paymentSum(foodPrice, drinkPrice, drinkCode, noDrink));
+    printf("Payment amount: %f\n", paymentSum(foodPrice, drinkPrice, drinkCode, noDrink));
     printf("-----------------\n");
 }
 
@@ -36,8 +37,8 @@ void printFormHeader(char* name) {
            "Food items:\n", name);
 }
 
-int paymentSum(int foodPrice, int drinkPrice, int drinkCode, int noDrink) {
-    int sum=0;
+double paymentSum(double foodPrice, double drinkPrice, int drinkCode, int noDrink) {
+    double sum=0;
     if(drinkCode != noDrink)
         sum=foodPrice+drinkPrice;
     else sum=foodPrice;
